file-structure refactoring/folder renaming . sim is now engine, env is now enviroment
