+++ index.html
@@ -9,7 +9,6 @@
   <body>
     <div id="boot" aria-live="polite">
       <div class="boot-card">
-        <h1 class="boot-title">Loading…</h1>
         <div class="boot-bar"><div id="bootProgressFill" class="boot-fill" style="width:0%"></div></div>
         <div class="boot-meta">
           <span id="bootPercent">0%</span>
