+++ src/engine/heart/curve.ts
