+++ src/engine/trails.ts
