+++ src/environment/lighting.ts
