--- src/sim/heart.ts
@@ -1,2 +1,0 @@
-export { HEART_CENTER, heartFrame, updateHeartFrame, heartLocalToWorld } from "@/sim/heart/frame";
-export { generateHeartPoints, generateHeartPointsVariable } from "@/sim/heart/sampling";
